+++ packages/core/src/agent/index.ts
@@ -927,7 +927,7 @@ export class Agent<
 
       const newMessages = messageList.get.input.v1() as CoreMessage[];
 
-      const processedMemoryMessages = memory.processMessages({
+      const processedMemoryMessages = await memory.processMessages({
         // these will be processed
         messages: messageList.get.remembered.v1() as CoreMessage[],
         // these are here for inspecting but shouldn't be returned by the processor
@@ -1880,7 +1880,7 @@ Message ${msg.threadId && msg.threadId !== threadObject.id ? 'from previous conv
             ?.map(m => m.content)
             ?.join(`\n`) ?? undefined;
 
-        const processedMemoryMessages = memory.processMessages({
+        const processedMemoryMessages = await memory.processMessages({
           // these will be processed
           messages: messageList.get.remembered.v1() as CoreMessage[],
           // these are here for inspecting but shouldn't be returned by the processor
@@ -2791,7 +2791,7 @@ Message ${msg.threadId && msg.threadId !== threadObject.id ? 'from previous conv
             ?.map(m => m.content)
             ?.join(`\n`) ?? undefined;
 
-        const processedMemoryMessages = memory.processMessages({
+        const processedMemoryMessages = await memory.processMessages({
           // these will be processed
           messages: messageList.get.remembered.v1() as CoreMessage[],
           // these are here for inspecting but shouldn't be returned by the processor
