+++ packages/memory/integration-tests/src/processors.test.ts
@@ -81,7 +81,7 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result = memory.processMessages({
+    const result = await memory.processMessages({
       messages: new MessageList({ threadId: thread.id, resourceId })
         .add(queryResult.uiMessages, 'memory')
         .get.all.core(),
@@ -111,7 +111,7 @@ describe('Memory with Processors', () => {
     });
     expect(allMessagesQuery.messages.length).toBe(20);
 
-    const allMessagesResult = memory.processMessages({
+    const allMessagesResult = await memory.processMessages({
       messages: new MessageList({ threadId: thread.id, resourceId })
         .add(allMessagesQuery.uiMessages, 'memory')
         .get.all.core(),
@@ -155,7 +155,7 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result = memory.processMessages({
+    const result = await memory.processMessages({
       messages: v2ToCoreMessages(queryResult.uiMessages),
       processors: [new ToolCallFilter({ exclude: ['weather'] })],
     });
@@ -171,7 +171,10 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result2 = memory.processMessages({ messages: v2ToCoreMessages(queryResult2.uiMessages), processors: [] });
+    const result2 = await memory.processMessages({
+      messages: v2ToCoreMessages(queryResult2.uiMessages),
+      processors: [],
+    });
     const messages2 = new MessageList({ threadId: thread.id, resourceId }).add(result2, 'response').get.all.v2();
     expect(new MessageList().add(messages2, 'memory').get.all.v2()).toHaveLength(messagesV2.length);
     expect(filterToolCallsByName(result2, 'weather')).toHaveLength(1);
@@ -184,7 +187,7 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result3 = memory.processMessages({
+    const result3 = await memory.processMessages({
       messages: v2ToCoreMessages(queryResult3.uiMessages),
       processors: [new ToolCallFilter({ exclude: ['weather', 'calculator'] })],
     });
@@ -199,7 +202,7 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result4 = memory.processMessages({
+    const result4 = await memory.processMessages({
       messages: v2ToCoreMessages(queryResult4.uiMessages),
       processors: [new ToolCallFilter()],
     });
@@ -234,7 +237,7 @@ describe('Memory with Processors', () => {
       threadId: thread.id,
       selectBy: { last: 20 },
     });
-    const result = memory.processMessages({
+    const result = await memory.processMessages({
       messages: v2ToCoreMessages(queryResult.uiMessages),
       processors: [new ToolCallFilter({ exclude: ['weather'] }), new TokenLimiter(250)],
     });
@@ -415,7 +418,7 @@ describe('Memory with Processors', () => {
 
     const list = new MessageList({ threadId }).add(queryResult.messagesV2, 'memory');
 
-    const baselineResult = memory.processMessages({
+    const baselineResult = await memory.processMessages({
       messages: list.get.remembered.core(),
       newMessages: list.get.input.core(),
       processors: [],
@@ -436,7 +439,7 @@ describe('Memory with Processors', () => {
       selectBy: { last: 20 },
     });
     const list2 = new MessageList({ threadId }).add(weatherQueryResult.messagesV2, 'memory');
-    const weatherFilteredResult = memory.processMessages({
+    const weatherFilteredResult = await memory.processMessages({
       messages: list2.get.all.core(),
       processors: [new ToolCallFilter({ exclude: ['get_weather'] })],
     });
@@ -457,7 +460,7 @@ describe('Memory with Processors', () => {
       selectBy: { last: 20 },
     });
     const list3 = new MessageList({ threadId }).add(tokenLimitQuery.messages, 'memory');
-    const tokenLimitedResult = memory.processMessages({
+    const tokenLimitedResult = await memory.processMessages({
       messages: list3.get.all.core(),
       processors: [new TokenLimiter(100)], // Small limit to only get a subset
     });
@@ -471,7 +474,7 @@ describe('Memory with Processors', () => {
       selectBy: { last: 20 },
     });
     const list4 = new MessageList({ threadId }).add(combinedQuery.messages, 'memory');
-    const combinedResult = memory.processMessages({
+    const combinedResult = await memory.processMessages({
       messages: list4.get.all.core(),
       processors: [new ToolCallFilter({ exclude: ['get_weather', 'calculator'] }), new TokenLimiter(500)],
     });
@@ -524,7 +527,7 @@ describe('Memory with Processors', () => {
     });
 
     // Retrieve the message (no TokenLimiter, just get the message back)
-    const result = memory.processMessages({
+    const result = await memory.processMessages({
       messages: v2ToCoreMessages(queryResult.uiMessages),
     });
